AccessRoomForm to module
